==> codigo/Scripts Finales/2_Evaluacion_de_Coberturas.R <==
# PAQUETES Y DATOS--------------------------------------------------------------
library(tidyverse)
library(sf)
sf::sf_use_s2(FALSE)
library(ggplot2)
library(patchwork)
library(lubridate)
library(geouy)


NatUY <- read_csv('datos/Tablas/Observaciones_27-10-22.csv')
Uruguay <- geouy::load_geouy("Dptos")
UY <- st_union(Uruguay) %>% st_cast()
deptosuy <- st_read("datos/Datos espaciales/ine_depto.shp")


# PROCESAMIENTO PREVIO DE DATOS-------------------------------------------------

## Convertir NatUY en un Simple Feature
NatUY_sf <- NatUY %>% 
  st_as_sf(coords = c("longitude", "latitude")) %>% 
  st_set_crs(4326) %>% 
  st_transform(32721)

## Filtrado de registros: GI + Nivel Especie
listado_especies <- NatUY_sf %>% 
  select(observed_on,quality_grade, place_admin1_name, 
         taxon_species_name,taxon_kingdom_name, taxon_phylum_name, 
         taxon_class_name,taxon_order_name, taxon_family_name, 
         taxon_genus_name,scientific_name) %>% 
  filter(quality_grade == "research" & !is.na(taxon_species_name) & 
           taxon_species_name!="")


saveRDS(NatUY_sf, "datos/Tablas/natuysf.rds")
saveRDS(listado_especies, "datos/Tablas/listado_especies.rds")


# ANALISIS COBERTURA ESPACIAL--------------------------------------------------

### Creamos una grilla para Uruguay
Grilla_UY <- st_make_grid(st_bbox(UY), 
                          cellsize = 25000, square = FALSE, 
                          crs= st_crs(Uruguay)) %>% 
  st_intersection(UY) %>% st_sf(gridID=1:length(.), geometry= .)


### Unimos la grilla con la tabla de especies filtrada 
NatUY_grilla <- st_join(Grilla_UY, listado_especies) %>% 
  group_by(gridID) %>% 
  summarise(registros=n(),
            especies=n_distinct(scientific_name)) %>% st_cast()


## Mapa cantidad de Registros
plot_registros <- ggplot() +
  geom_sf(data=NatUY_grilla, aes(fill=log10(registros)),show.legend = T) +
  labs(fill="Acumulación de Registros") +
  scale_fill_gradient2(high = "#034E7B", low = "#D0D1E6") +
  theme(legend.text = element_text(size = 12)) +
  theme_bw()


## Mapa cantidad de especies registradas
plot_riqueza <- ggplot() +
  geom_sf(data=NatUY_grilla, aes(fill=log10(especies)),show.legend = T) + 
  labs(fill="Riqueza de especies") + 
  scale_fill_gradient2(high = "#990000", low = "#FDD49E") + 
  theme(legend.text = element_text(size = 12)) + 
  theme_bw()

plot_registros + geom_sf(data= deptosuy, color = "black", fill= NA) + 
  plot_riqueza + geom_sf(data= deptosuy, color = "black", fill= NA)


## Cobertura por departamentos
listado_especies %>% st_drop_geometry() %>% 
  group_by(place_admin1_name) %>% 
  summarise(registros=n(),especies=n_distinct(scientific_name)) %>% 
  arrange(desc(registros))


# ANALISIS COBERTURA TEMPORAL---------------------------------------------------

## Cobertura temporal de los registros
registros_por_año <- listado_especies %>% 
  st_drop_geometry() %>% 
  filter(year(observed_on)>=2010) %>% 
  add_count(year=year(observed_on), 
            name = "registros_por_año") %>% 
  ggplot(., aes(x=observed_on, y=registros_por_año)) +
  geom_line(size=0.5, show.legend = FALSE) +
  scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
  scale_y_continuous(breaks = seq(0, 12500, by = 2500)) + 
  theme_bw() +
  labs(x="", y= "Numero de registros")

## Cobertura temporal por estaciones
registros_por_estaciones <- listado_especies %>% 
  st_drop_geometry() %>% 
  filter(year(observed_on)>=2010) %>% 
  mutate(observed_on=as_date(observed_on)) %>% 
  mutate(season=lubridate::quarter(observed_on)) %>% 
  mutate(season=ifelse(season==1, 'verano', 
                       ifelse(season==2, 'otoño', 
                              ifelse(season==3, 'invierno', 'primavera')))) %>% 
  add_count(season, name = "registros_por_estacion") %>% 
  mutate(season=factor(season, 
                       levels = c("otoño", "invierno","primavera", "verano"))) %>% 
  ggplot(aes(x=season, y=registros_por_estacion)) +
  geom_segment(aes(x=season,xend=season,y=0, 
                   yend= registros_por_estacion, col=season),
               show.legend = FALSE) +
  geom_point(aes(col=season),show.legend = FALSE) +
  scale_y_continuous(breaks = seq(0, 12500, by = 2500)) +
  theme_bw() +
  labs(x="", y= "Numero de registros")

registros_por_año | registros_por_estaciones


## Registro temporal de los principales Reinos
Temporal_Reinos <- listado_especies %>% st_drop_geometry() %>%
  filter(year(observed_on)>=2010) %>% 
  filter(!is.na(taxon_phylum_name)) %>% 
  filter(taxon_kingdom_name=='Plantae' | taxon_kingdom_name=='Animalia' | 
           taxon_kingdom_name=='Fungi') %>% 
  group_by(year(observed_on), taxon_kingdom_name, taxon_phylum_name) %>% 
  count() %>% 
  ggplot(aes(x=`year(observed_on)`, y= n, color=taxon_phylum_name), 
         group = taxon_kingdom_name) + 
  facet_wrap(taxon_kingdom_name~.,scales = "free",drop=TRUE, 
             ncol = 1, nrow = 3) +
  geom_line(size=1, show.legend = FALSE) + geom_point() +
  labs(x='Años', y='Registros', color = "Filos") +
  theme_bw()


### Registro temporal de Animalia por clases
Temporal_A <- listado_especies %>% st_drop_geometry() %>%
  filter(year(observed_on)>=2010) %>% 
  filter(!is.na(taxon_phylum_name)) %>% 
  filter(taxon_kingdom_name=='Animalia') %>% 
  group_by(year(observed_on), taxon_kingdom_name, taxon_phylum_name,
           taxon_class_name) %>% count() %>%
  ggplot(aes(x=`year(observed_on)`, y= n, color=taxon_phylum_name,)) + 
  geom_line(size=1, show.legend = TRUE) + geom_point() +
  scale_x_continuous(breaks= scales::pretty_breaks()) +
  facet_wrap(taxon_class_name~.,scales = "free",drop=TRUE, nrow = 3) +
  labs(x='Años', y='Registros', color = 'filo') +
  theme_bw()


### Registro temporal del Reino Plantae por clases
Temporal_P <- listado_especies %>% st_drop_geometry() %>%
  filter(year(observed_on)>=2010) %>% 
  filter(!is.na(taxon_phylum_name)) %>% 
  filter(taxon_kingdom_name=='Plantae') %>% 
  group_by(year(observed_on), taxon_kingdom_name, taxon_phylum_name, 
           taxon_class_name) %>% count() %>% 
  ggplot(aes(x=`year(observed_on)`, y= n, color=taxon_phylum_name,)) + 
  geom_line(size=1, show.legend = TRUE) + geom_point() +
  scale_x_continuous(breaks= scales::pretty_breaks()) +
  facet_wrap(taxon_class_name~.,scales = "free",drop=TRUE) +
  labs(x='Años', y='Registros', color = 'Filo') +
  theme_bw()


### Registro temporal del reino Fungi por clases
Temporal_F <- listado_especies %>% st_drop_geometry() %>%
  filter(year(observed_on)>=2010) %>% 
  filter(!is.na(taxon_phylum_name)) %>% 
  filter(taxon_kingdom_name=='Fungi') %>% 
  group_by(year(observed_on), taxon_kingdom_name, taxon_phylum_name,
           taxon_class_name) %>% count() %>% 
  ggplot(aes(x=`year(observed_on)`, y= n, color=taxon_phylum_name,)) + 
  geom_line(size=1, show.legend = TRUE) + geom_point() +
  scale_x_continuous(breaks= scales::pretty_breaks()) +
  facet_wrap(taxon_class_name~.,scales = "free",drop=TRUE) +
  labs(x='Años', y='Registros', color = 'filo') +
  theme_bw()

Temporal_A / (Temporal_P | Temporal_F)


# ANALISIS COBERTURA TAXONÓMICA-------------------------------------------------

## 5 FILOS MAS REGISTRADOS
Taxon_Filos <- listado_especies %>% 
  filter(taxon_kingdom_name=='Animalia' | 
           taxon_kingdom_name=='Fungi' | 
           taxon_kingdom_name=='Plantae') %>% 
  group_by(taxon_kingdom_name, taxon_phylum_name) %>% 
  count() %>% arrange(desc(n)) %>% 
  filter(!is.na(taxon_phylum_name)) %>% head(6) %>%  
  ggplot(., aes(x=n, y= fct_reorder(taxon_phylum_name,n), 
                fill=taxon_kingdom_name)) +
  geom_bar(stat = "identity", show.legend = TRUE) +
  labs(title="Filos mas registrados",x='Registros', y= '', fill = 'Reino') +
  theme_bw() + scale_x_continuous()


## 10 CLASES MAS REGISTRADAS
Taxon_Clases <- listado_especies %>%  
  filter(taxon_kingdom_name=='Animalia' | taxon_kingdom_name=='Fungi' | 
           taxon_kingdom_name=='Plantae') %>% 
  group_by(taxon_kingdom_name, taxon_phylum_name, taxon_class_name) %>% 
  count() %>% arrange(desc(n)) %>% head(10) %>% 
  filter(!is.na(taxon_class_name)) %>% 
  ggplot(.,aes(x=n, y=fct_reorder(taxon_class_name,n), 
               fill=taxon_kingdom_name)) +
  geom_bar(stat = "identity", show.legend = T) +
  labs(title="Clases mas registradas", 
       x='Registros', y= '', fill = 'Reino') + theme_bw() +
  scale_x_continuous()
